refactor: Network Provider
