+++ src/components/Tips/TokenItem.tsx
@@ -5,7 +5,7 @@ import { router, TipsRoutePath } from '@/components/Tips/tipsModalRouter.js';
 import { TokenIcon } from '@/components/Tips/TokenIcon.js';
 import { NetworkType } from '@/constants/enum.js';
 import { classNames } from '@/helpers/classNames.js';
-import { resolveNetworkProvider } from '@/helpers/resolveTokenTransfer.js';
+import { switchEthereumChain } from '@/helpers/switchEthereumChain.js';
 import { TipsContext } from '@/hooks/useTipsContext.js';
 import type { Token } from '@/providers/types/Transfer.js';
 
@@ -18,10 +18,7 @@ export function TokenItem({ token }: TokenItemProps) {
 
     const handleSelectToken = async (token: Token) => {
         if (recipient?.networkType === NetworkType.Ethereum) {
-            const network = resolveNetworkProvider(recipient.networkType);
-            if (token.chainId !== network.getChainId()) {
-                await network.switchChain(token.chainId);
-            }
+            await switchEthereumChain(token.chainId);
         }
         update((prev) => ({ ...prev, token, amount: token.id !== selectedToken?.id ? '' : prev.amount }));
         router.navigate({ to: TipsRoutePath.TIPS, replace: true });
