+++ src/providers/ethereum/Network.ts
@@ -1,9 +1,9 @@
 import { EVMExplorerResolver } from '@masknet/web3-providers';
 import type { ChainId } from '@masknet/web3-shared-evm';
-import { getAccount, getChainId, switchChain } from '@wagmi/core';
+import { getAccount, getChainId } from '@wagmi/core';
 import { type Address, type Hash } from 'viem';
 
-import { chains, config } from '@/configs/wagmiClient.js';
+import { config } from '@/configs/wagmiClient.js';
 import { NotImplementedError } from '@/constants/error.js';
 import type { NetworkProvider } from '@/providers/types/Network.js';
 
@@ -18,21 +18,6 @@ class Provider implements NetworkProvider<ChainId, Address, Hash> {
             throw new Error('Wallet not connected');
         }
         return account.address;
-    }
-
-    async switchChain(chainId: ChainId): Promise<void> {
-        const chain = chains.find((chain) => chain.id === chainId);
-        await switchChain(config, {
-            chainId,
-            addEthereumChainParameter: chain
-                ? {
-                      chainName: chain.name,
-                      nativeCurrency: chain.nativeCurrency,
-                      rpcUrls: chain.rpcUrls.default.http,
-                      blockExplorerUrls: [chain.blockExplorers.default.url],
-                  }
-                : undefined,
-        });
     }
 
     getChainId(): ChainId {
