+++ src/providers/ethereum/Transfer.ts
@@ -7,6 +7,7 @@ import { type Address, erc20Abi, type Hash, parseUnits } from 'viem';
 import { config } from '@/configs/wagmiClient.js';
 import { formatBalance } from '@/helpers/formatBalance.js';
 import { isZero, multipliedBy } from '@/helpers/number.js';
+import { switchEthereumChain } from '@/helpers/switchEthereumChain.js';
 import { getAvailableBalance } from '@/providers/ethereum/getAvailableBalance.js';
 import { getDefaultGas } from '@/providers/ethereum/getDefaultGas.js';
 import { isNativeToken } from '@/providers/ethereum/isNativeToken.js';
@@ -17,15 +18,12 @@ class Provider implements TransferProvider<ChainId, Address, Hash> {
     async transfer(options: TransactionOptions<ChainId, Address>): Promise<Address> {
         const { token } = options;
         if (token.chainId !== EthereumNetwork.getChainId()) {
-            await EthereumNetwork.switchChain(token.chainId);
+            await switchEthereumChain(token.chainId);
         }
 
-        let hash: Address;
-        if (this.isNativeToken(token)) {
-            hash = await this.transferNative(options);
-        } else {
-            hash = await this.transferContract({ ...options, token });
-        }
+        const hash = this.isNativeToken(token)
+            ? await this.transferNative(options)
+            : await this.transferContract({ ...options, token });
 
         await this.waitForTransaction(hash);
         return hash;
