+++ src/providers/solana/Network.ts
@@ -1,7 +1,6 @@
 import { ChainId } from '@masknet/web3-shared-solana';
 import { WalletNotConnectedError } from '@solana/wallet-adapter-base';
 
-import { NotImplementedError } from '@/constants/error.js';
 // eslint-disable-next-line no-restricted-imports
 import { SolanaExplorerResolver } from '@/maskbook/packages/web3-providers/src/Web3/Solana/apis/ResolverAPI.js';
 import { resolveWalletAdapter } from '@/providers/solana/resolveWalletAdapter.js';
@@ -22,10 +21,6 @@ class Provider implements NetworkProvider<ChainId> {
         if (!adapter.publicKey) throw new WalletNotConnectedError();
 
         return adapter.publicKey.toBase58();
-    }
-
-    switchChain(chainId: ChainId): Promise<void> {
-        throw new NotImplementedError();
     }
 
     getChainId(): ChainId {
